MatrixDesign
What is MatrixDesign?
MatrixDesign basically is looping process that will print a square with 4 divided side. The side separator will be in a triangle.

Who involve in this MatrixDesign?
This program originally i got from one of the test i have done. But to make it a little different i do some enhancement on the question. The person who involve in this program is only ME and no AI help during the process of creation.

How long i made this program:
I build this program about 3-4 hours